pyproject.toml
[project]
name = "paper-audit-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.110",
    "uvicorn[standard]>=0.29",
    "chromadb>=0.4.24",           # 替代 Milvus，纯本地向量库
    "httpx>=0.27",                # 替代 gRPC 客户端，异步 HTTP
    "langgraph>=0.0.30",          # 工作流
    "pydantic-settings>=2.2",     # .env 配置管理
    "aiosqlite>=0.20",            # 异步 SQLite（任务队列）
    "python-multipart>=0.0.9",    # 文件上传支持
    "PyMuPDF>=1.24",              # PDF 报告生成（fitz）
    "python-docx>=1.1",           # Python 端 Word 辅助处理（可选）
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.22",
    "ruff>=0.12.0",
    "black>=23.1.0",
]